requirements.txt
fastapi[all]
sqlalchemy
psycopg2-binary
alembic

pytest-alembic
docker
faker